move ai and wave listener to different folder
